+++ tests/test_read.py
@@ -6,7 +6,7 @@ from six.moves import xrange as _xrange  # pylint: disable=F0401
 from . import BaseSystemTest, is_number
 from dynamo3 import STRING, NUMBER, DynamoKey, LocalIndex, GlobalIndex, TOTAL
 from dynamo3.result import Result, GetResultSet
-from mock import MagicMock
+from mock import MagicMock, ANY
 
 
 class TestQuery(BaseSystemTest):
@@ -459,6 +459,11 @@ class TestQuery2(BaseSystemTest):
                                      id='a', a='a', b='a')
         self.assertItemsEqual(list(results), [a, b])
 
+    def test_dry_run(self):
+        """ dry_run=True """
+        ret = self.dynamo.query2('foobar', 'id = :id', id='a', dry_run=True)
+        self.assertEqual(ret, ('Query', ANY))
+
 
 class TestScan(BaseSystemTest):
 
@@ -887,6 +892,11 @@ class TestScan2(BaseSystemTest):
                               [{'id': 'a'}, {'id': 'b'},
                                {'id': 'c'}, {'id': 'd'}])
 
+    def test_dry_run(self):
+        """ dry_run=True """
+        ret = self.dynamo.scan2('foobar', dry_run=True)
+        self.assertEqual(ret, ('Scan', ANY))
+
 
 class TestBatchGet(BaseSystemTest):
 
@@ -976,6 +986,12 @@ class TestBatchGet(BaseSystemTest):
         self.assertEqual(rs.indexes, {'l-index': 1})
         self.assertEqual(rs.global_indexes, {'g-index': 1})
 
+    def test_dry_run(self):
+        """ dry_run=True """
+        keys = [{'id': 'a'}, {'id': 'b'}]
+        ret = self.dynamo.batch_get('foobar', keys, dry_run=True)
+        self.assertEqual(ret, ('BatchGetItem', ANY))
+
 
 class TestGetItem(BaseSystemTest):
 
@@ -1019,3 +1035,61 @@ class TestGetItem(BaseSystemTest):
         response = {'Item': self.dynamo.dynamizer.encode_keys(d)}
         result = Result(self.dynamo.dynamizer, response, 'Item')
         self.assertNotEqual(repr(result), repr(d))
+
+
+class TestGetItem2(BaseSystemTest):
+
+    """ Tests for new GetItem API """
+
+    def make_table(self):
+        """ Convenience method for making a table """
+        hash_key = DynamoKey('id')
+        self.dynamo.create_table('foobar', hash_key=hash_key)
+
+    def test_get(self):
+        """ Can fetch an item by the primary key """
+        self.make_table()
+        item = {'id': 'a', 'foo': 'bar'}
+        self.dynamo.put_item('foobar', item)
+        ret = self.dynamo.get_item2('foobar', {'id': 'a'})
+        self.assertEqual(ret, item)
+
+    def test_attribute(self):
+        """ Can fetch only certain attributes """
+        self.make_table()
+        item = {'id': 'a', 'foo': 'bar'}
+        self.dynamo.put_item('foobar', item)
+        ret = self.dynamo.get_item2('foobar', {'id': 'a'}, attributes='id')
+        self.assertEqual(ret, {'id': 'a'})
+
+    def test_attribute_alias(self):
+        """ GetItem with ExpressionAttributeNames """
+        self.make_table()
+        item = {'id': 'a', 'foo': 'bar'}
+        self.dynamo.put_item('foobar', item)
+        ret = self.dynamo.get_item2('foobar', {'id': 'a'}, attributes=['#i'],
+                                    alias={'#i': 'id'})
+        self.assertEqual(ret, {'id': 'a'})
+
+    def test_capacity(self):
+        """ Can return the consumed capacity as well """
+        self.make_table()
+        self.dynamo.put_item('foobar', {'id': 'a'})
+        ret = self.dynamo.get_item2('foobar', {'id': 'a'},
+                                    return_capacity=TOTAL)
+        self.assertTrue(is_number(ret.capacity))
+        self.assertTrue(is_number(ret.table_capacity))
+        self.assertTrue(isinstance(ret.indexes, dict))
+        self.assertTrue(isinstance(ret.global_indexes, dict))
+
+    def test_result_repr(self):
+        """ Result repr should not be the same as a dict """
+        d = {'a': 'b'}
+        response = {'Item': self.dynamo.dynamizer.encode_keys(d)}
+        result = Result(self.dynamo.dynamizer, response, 'Item')
+        self.assertNotEqual(repr(result), repr(d))
+
+    def test_dry_run(self):
+        """ dry_run=True """
+        ret = self.dynamo.get_item2('foobar', {'id': 'a'}, dry_run=True)
+        self.assertEqual(ret, ('GetItem', ANY))